+++ src/album.rs
@@ -1,5 +1,7 @@
 use crate::genius::GeniusAlbumResponse;
+use crate::settings::{Settings, settings_from_req};
 use crate::utils;
+use actix_web::HttpRequest;
 use actix_web::{get, web, Responder, Result};
 use askama::Template;
 use serde::Deserialize;
@@ -11,6 +13,7 @@ use crate::templates::template;
 #[derive(Template)]
 #[template(path = "album.html")]
 struct AlbumTemplate {
+    settings: Settings,
     album: GeniusAlbum,
 }
 
@@ -20,7 +23,7 @@ pub struct AlbumQuery {
 }
 
 #[get("/album")]
-pub async fn album(info: web::Query<AlbumQuery>) -> Result<impl Responder> {
+pub async fn album(req: HttpRequest, info: web::Query<AlbumQuery>) -> Result<impl Responder> {
     let album_res = GeniusApi::global()
         .extract_data::<GeniusAlbumResponse>(&utils::ensure_path_prefix("albums", &info.path))
         .await?;
@@ -28,5 +31,5 @@ pub async fn album(info: web::Query<AlbumQuery>) -> Result<impl Responder> {
 
     album.tracks = Some(GeniusApi::global().get_album_tracks(album.id).await?);
 
-    Ok(template(AlbumTemplate { album }))
+    Ok(template(AlbumTemplate { settings: settings_from_req(&req), album }))
 }